pyproject.toml
[project]
name = "dtc-ml-midterm-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

authors = [
    { name = "Leonel Aliaga", email = "leonel.aliaga.v@gmail.com"}
    ]

dependencies = [
    "fastapi>=0.121.2",
    "hydra-core>=1.3.2",
    "keras-tuner>=1.4.8",
    "lightgbm>=4.6.0",
    "numpy>=2.3.4",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
]
